created second component for navigating
